# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Do it - Productivity PWA API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface Task {
  id: string;
  userId: string;
  titulo: string;
  /** @nullable */
  descripcion?: string | null;
  /** @nullable */
  fechaVencimiento?: string | null;
  completada: boolean;
  createdAt: string;
}

export interface TaskInput {
  /** @minLength 1 */
  titulo: string;
  /** @nullable */
  descripcion?: string | null;
  /** @nullable */
  fechaVencimiento?: string | null;
  userId: string;
}

export interface TaskUpdate {
  titulo?: string;
  /** @nullable */
  descripcion?: string | null;
  /** @nullable */
  fechaVencimiento?: string | null;
  completada?: boolean;
}

export interface TaskStats {
  total: number;
  completed: number;
  pending: number;
  completedToday: number;
}

export interface Habit {
  id: string;
  userId: string;
  nombre: string;
  frecuencia: string;
  createdAt: string;
}

export interface HabitWithLogs {
  id: string;
  userId: string;
  nombre: string;
  frecuencia: string;
  createdAt: string;
  /** Array of ISO date strings where habit was completed (last 7 days) */
  logs: string[];
}

export interface HabitInput {
  /** @minLength 1 */
  nombre: string;
  frecuencia?: string;
  userId: string;
}

export interface HabitLog {
  id: string;
  habitId: string;
  fechaCompletado: string;
}

export interface TranscribeInput {
  userId: string;
  /** Base64-encoded audio data */
  audioBase64: string;
  /** MIME type of the audio (e.g. audio/webm) */
  mimeType: string;
}

export type GetTasksParams = {
completed?: boolean;
};