# lib/api-zod/src/generated/types/habitInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Do it - Productivity PWA API
 * OpenAPI spec version: 0.1.0
 */

export interface HabitInput {
  /** @minLength 1 */
  nombre: string;
  frecuencia?: string;
  userId: string;
}