# lib/api-zod/src/generated/types/habitLog.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Do it - Productivity PWA API
 * OpenAPI spec version: 0.1.0
 */

export interface HabitLog {
  id: string;
  habitId: string;
  fechaCompletado: string;
}