# lib/api-zod/src/generated/types/habitWithLogs.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Do it - Productivity PWA API
 * OpenAPI spec version: 0.1.0
 */

export interface HabitWithLogs {
  id: string;
  userId: string;
  nombre: string;
  frecuencia: string;
  createdAt: string;
  /** Array of ISO date strings where habit was completed (last 7 days) */
  logs: string[];
}