# lib/api-zod/src/generated/types/taskStats.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Do it - Productivity PWA API
 * OpenAPI spec version: 0.1.0
 */

export interface TaskStats {
  total: number;
  completed: number;
  pending: number;
  completedToday: number;
}