# lib/api-zod/src/generated/types/taskUpdate.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Do it - Productivity PWA API
 * OpenAPI spec version: 0.1.0
 */

export interface TaskUpdate {
  titulo?: string;
  /** @nullable */
  descripcion?: string | null;
  /** @nullable */
  fechaVencimiento?: string | null;
  completada?: boolean;
}