# lib/api-zod/src/generated/types/transcribeInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Do it - Productivity PWA API
 * OpenAPI spec version: 0.1.0
 */

export interface TranscribeInput {
  userId: string;
  /** Base64-encoded audio data */
  audioBase64: string;
  /** MIME type of the audio (e.g. audio/webm) */
  mimeType: string;
}